+++ src/main/java/com/bobeat/backend/domain/store/util/PgVectorUtils.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 public class PgVectorUtils {
-    public static String toLiteral(List<Double> vector) {
+    public static String toLiteral(List<Float> vector) {
         return "[" + vector.stream()
                 .map(d -> String.format("%.6f", d))
                 .collect(Collectors.joining(",")) + "]";
